Test decoupling typing aliases
